script: Move all focus-related code to `components/script/document/focus.rs` and create `DocumentFocusHandler`
This continues the work to split up the large DOM structs. In this case
we create a helper struct to store focus-related state much like
`DocumentEventHandler`.
